requirements/requirements_0.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
backcall==0.2.0
bitsandbytes==0.45.0
blessed==1.20.0
braceexpand==0.1.7
cachetools==5.5.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.17.1
chainer==7.8.1
channelnorm-cuda==0.0.0
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.2
colorama==0.4.6
contourpy==1.1.1
correlation-cuda==0.0.0
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
Cython==0.29
data==0.4
decorator==4.4.2
decord==0.6.0
deepspeed==0.9.2
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.7.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
ffmpeg==1.4
ffmpy==0.4.0
filelock==3.14.0
fonttools==4.53.1
frozenlist==1.5.0
fschat==0.1.10
fsspec==2024.10.0
ftfy==6.2.3
funcsigs==1.0.2
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
gradio==3.23.0
gradio_client==0.0.8
grpcio==1.67.1
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.6.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
latex2mathml==3.77.0
lazy_loader==0.4
linkify-it-py==2.0.3
lit==18.1.8
Markdown==3.7
markdown-it-py==2.2.0
markdown2==2.5.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdtex2html==1.3.0
mdurl==0.1.2
mmcv-full==1.3.1
mmdet==2.11.0
mmpycocotools==12.0.3
model-index==0.1.11
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
mypy-extensions==1.0.0
narwhals==1.12.1
networkx==3.1
ninja==1.11.1.1
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.52.2
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
orjson==3.10.10
oss2==2.17.0
packaging==24.1
pandas==2.0.3
parameterized==0.9.0
parso==0.8.4
peft==0.3.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==7.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
portalocker==2.10.1
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.7
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytorchvideo==0.1.5
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.35.1
regex==2024.9.11
requests==2.28.2
requests-oauthlib==2.0.0
resample2d-cuda==0.0.0
rich==13.4.2
rpds-py==0.20.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.6.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.16.0
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.2
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.11
tokenizers==0.20.1
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.4.1
tqdm==4.65.2
traitlets==5.14.3
transformers==4.46.1
triton==2.0.0
typed-argument-parser==1.10.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==1.26.20
uvicorn==0.32.0
visdom==0.2.4
wandb==0.18.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
webdataset==0.2.100
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
yacs==0.1.8
yapf==0.40.2
yarl==1.15.2
zipp==3.20.2
